third_party/blink/web_tests/fast/dom/Window/open-after-frame-detached.html

<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script>
var childWindow;
function runTest()
{
    description("Tests that calling window.open on the contentWindow of a detached frame does not crash or open a new window. You may need to disable popup blocking for this test case.");
    if (window.testRunner)
        testRunner.setPopupBlockingEnabled(false);
    var ifr = document.createElement("iframe");
    ifr.onload = function ()
    {
        childWindow = ifr.contentWindow;
        ifr.parentNode.removeChild(ifr);
        shouldBeNull("childWindow.open('about:blank')");
    }
    document.body.appendChild(ifr);
}
</script>
</head>
<body onload="runTest()">
</body>
</html>
